added reusing sbvfs to e2e test
